Add files via upload
These are header/resource/source files dedicated to a pingpong clone I created in Visual Studio 2022.  I created the project using the "Windows Desktop Application" open.  This generates several files which include a template api, menu, and image.  90% of my code modifications, edits, and additions are located in the PingPongClone.cpp file.  I'm including some of the resource/header files so that someone new can follow the steps that were taken.  Code is commented and should be accurate.  Places that could use improvement are also marked in some locations.